% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrace.mcp.spearman.sampling.R
\name{dtrace.mcp.spearman.sampling}
\alias{dtrace.mcp.spearman.sampling}
\title{Estimate genetic network using MCP-penalized D-trace loss function and Spearman's rho correlation matrix}
\usage{
dtrace.mcp.spearman.sampling(
  BETA,
  Rnoise,
  lamvec = c(1:20)/100,
  max.eps = 0.01,
  max.iter = 25,
  rho = 0.05,
  mineig = 0.01,
  subtime = 100,
  subfrac = 0.5,
  subthres = 0.95,
  alpha = 0,
  bic.factor = 0.5,
  reliability.thres = 0.8,
  PenaMatrix = "none"
)
}
\arguments{
\item{BETA}{A matrix of dimensions mxp, where m is the number of variants and p is the number of exposures. It represents effect size estimates or Z-scores.}

\item{Rnoise}{The covariance matrix of estimation errors in BETA. If BETA contains Z-scores, Rnoise should be a correlation matrix.}

\item{lamvec}{A vector of tuning parameters for MCP. Default is c(1:20)/100.}

\item{max.eps}{The threshold for stopping the algorithm. Default is 0.01.}

\item{max.iter}{The maximum number of iterations for the algorithm. Since ADMM converges quickly but not with high precision, this should not be too large. Default is 25.}

\item{rho}{The penalty parameter for the ADMM algorithm. Default is 0.05.}

\item{mineig}{The minimum matrixEigenvalue for the precision matrix estimate. Default is 0.01.}

\item{subtime}{The number of times to resample. Default is 100.}

\item{subfrac}{The fraction of the data to resample each time. Default is 0.5.}

\item{subthres}{The threshold for stability selection. Default is 0.95.}

\item{alpha}{The additional parameter for MCP + alpha * Ridge. Default is 0.}

\item{bic.factor}{An additional penalty on cross-validation error. Default is 0.5.}

\item{reliability.thres}{The threshold for rescaling Rnoise to ensure that (var(BETAj)-Rnoisejj)/var(BETAj) is greater than this value. Default is 0.8.}

\item{PenaMatrix}{A penalty weight matrix to rescale the tuning parameter in MCP. Default is a matrix of ones.}
}
\value{
A list containing:
  - Theta: The estimated precision matrix.
  - Pvalue: P-values for the entries in Thetabased on subsampling.
  - K: Stability selection matrix indicating the proportion of subsamples where each entry in Theta was non-zero.
  - cv.error: The cross-validation error for each lambda in lamvec.
  - R: The Spearman rank correlation matrix of BETA.
  - ThetaSE: The standard error of Theta based on subsampling.
}
\description{
The \code{entropy.mcp.spearman.sampling} function estimates the genetic network by applying a penalized D-trace loss function using MCP (Minimax Concave Penalty) and Spearman's rho correlation matrix. It's designed to work with genetic data, specifically effect size estimates or Z-scores for various exposures across multiple variants.
}
\details{
The function performs subsampling and uses the MCP method to estimate a sparse precision matrix (inverse covariance matrix) representing the genetic network. It uses Spearman's rank correlation to handle non-linear relationships and applies stability selection to enhance the reliability of the selected network.
}
\examples{
# Assuming BETA and Rnoise are already defined:
result <- entropy.mcp.spearman.sampling(BETA, Rnoise)

}
