% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.R
\name{egg}
\alias{egg}
\title{egg}
\usage{
egg(stroke, egg = "", imegg, fade = 500, timeout = 5000,
  elementId = NULL)
}
\arguments{
\item{stroke}{keys that will trigger the egg.}

\item{egg}{message that will appear.}

\item{imegg}{image that appears.}

\item{fade}{in milliseconds.}

\item{timeout}{timeout animation in milliseconds.}

\item{elementId}{use to override element id.}
}
\description{
Add an eater egg to your document
}
\details{
You don't need to know the special Javascript key codes, as you can use
plain English to define your egg (but you can use key codes if you want to). i.e.:
\code{up} = up arrow on keyboard.
}
\examples{
\dontrun{
egg(c("r", "s", "t", "a", "t", "s"), "MY EASTER EGG APPEARS",
  imegg = "https://upload.wikimedia.org/wikipedia/commons/5/54/Bg-easter-eggs.jpg")
}

}

