% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_frame.r
\name{ggarrange}
\alias{ggarrange}
\title{gtable_frame}
\usage{
ggarrange(..., plots = list(...), nrow = NULL, ncol = NULL,
  widths = NULL, heights = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plots}{list of ggplots}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{widths}{list of requested widths}

\item{heights}{list of requested heights}
}
\value{
gtable of aligned plots
}
\description{
gtable_frame
}
\examples{
library(grid)
library(gtable)
p1 <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) +
  geom_point() 
p2 <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) +
  geom_point() + facet_wrap( ~ cyl, ncol=2, scales = "free") +
  guides(colour="none") +
  theme()
grid.newpage()
grid.draw(ggarrange(p1, p2, widths = lapply(c(1,2), unit, "null"), 
                    heights=list(unit(1,"null"))))
}

