% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggggr.r
\name{ggggr}
\alias{ggggr}
\title{symmetrise_scale}
\usage{
ggggr(g, nmax = 100)
}
\arguments{
\item{p}{ggplot2}

\item{axis}{axis}
}
\value{
plot with scales adjusted
}
\description{
symmetrise_scale
}
\examples{
library(ggplot2)
p1 <- qplot(mpg, wt, data=mtcars, colour=cyl) + facet_wrap(~carb, nrow=1, scales="free")
symmetrise_scale(p1, "y")
}
