% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel_size.r
\name{set_panel_size}
\alias{set_panel_size}
\title{set_panel_size}
\usage{
set_panel_size(p = NULL, g = ggplotGrob(p), file = NULL,
  margin = unit(1, "mm"), width = unit(4, "cm"), height = unit(4, "cm"))
}
\arguments{
\item{p}{ggplot2}

\item{g}{gtable}

\item{file}{optional output filename}

\item{margin}{grid unit}

\item{width}{grid unit, requested panel width}

\item{height}{grid unit, requested panel height}
}
\value{
gtable with fixed panel sizes
}
\description{
set_panel_size
}

