% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric_scale.r
\name{symmetric_range}
\alias{symmetric_range}
\title{symmetric_range}
\usage{
symmetric_range(range)
}
\arguments{
\item{range}{range of the data}
}
\value{
symmetric range
}
\description{
Function to ensure that a position scale is symmetric about 0
}
\examples{
library(ggplot2)
ggplot(mpg, aes(cty, hwy)) + 
geom_point() +
  scale_x_continuous(limits = symmetric_range)
}
