% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrise_scale.r
\name{symmetrise_scale}
\alias{symmetrise_scale}
\title{symmetrise_scale}
\usage{
symmetrise_scale(p, axis = "x")
}
\arguments{
\item{p}{ggplot2}

\item{axis}{axis}
}
\value{
plot with scales adjusted
}
\description{
Adds a blank layer to a ggplot to extend one axis symmetrically about 0 for all facets.
}
\examples{
library(ggplot2)
p1 <- qplot(mpg, wt, data=mtcars, colour=cyl) + facet_wrap(~carb, nrow=1, scales='free')
symmetrise_scale(p1, 'y')
}
