% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{theme_presentation}
\alias{theme_presentation}
\title{Theme with minimalistic (and opinionated) defaults suitable for presentation}
\usage{
theme_presentation(base_size = 24, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme with minimalistic (and opinionated) defaults suitable for presentation
}
