% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_frame.r
\name{gtable_frame}
\alias{gtable_frame}
\title{gtable_frame}
\usage{
gtable_frame(g, width = unit(1, "null"), height = unit(1, "null"))
}
\arguments{
\item{g}{gtable}

\item{width}{requested width}

\item{height}{requested height}
}
\value{
3x3 gtable wrapping the plot
}
\description{
gtable_frame
}

