% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_article.r
\name{theme_article}
\alias{theme_article}
\title{Theme with minimalistic (and opinionated) defaults suitable for publication}
\usage{
theme_article(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme based on ggthemes::theme_chew
}
\examples{
library(ggplot2)
ggplot() + theme_article()
}
